+++ src/features/repo-settings/RepoSettingsDialog.tsx
@@ -50,6 +50,7 @@ import type { HookDelivery, Webhook, WebhookInput } from "@/lib/git/types";
 import { formatRelativeTime } from "@/lib/time";
 import { toastError } from "@/lib/toast";
 import { cn } from "@/lib/utils";
+import { CollaboratorsSection } from "./CollaboratorsSection";
 import { FundingSection } from "./FundingSection";
 import { GeneralSettingsSection } from "./GeneralSettingsSection";
 import { SecretsSection } from "./SecretsSection";
@@ -106,6 +107,7 @@ export function RepoSettingsDialog({
         <Tabs defaultValue="general" className="flex min-h-0 min-w-0 flex-col">
           <TabsList>
             <TabsTrigger value="general">General</TabsTrigger>
+            <TabsTrigger value="access">Access</TabsTrigger>
             <TabsTrigger value="sponsor">Sponsor</TabsTrigger>
             <TabsTrigger value="secrets">Secrets &amp; variables</TabsTrigger>
             <TabsTrigger value="webhooks">Webhooks</TabsTrigger>
@@ -115,6 +117,12 @@ export function RepoSettingsDialog({
             className="min-h-0 min-w-0 overflow-y-auto pr-1"
           >
             <GeneralSettingsSection repoPath={repoPath} open={open} />
+          </TabsContent>
+          <TabsContent
+            value="access"
+            className="min-h-0 min-w-0 overflow-y-auto pr-1"
+          >
+            <CollaboratorsSection repoPath={repoPath} open={open} />
           </TabsContent>
           <TabsContent
             value="sponsor"